refactor(voice-rbac): break circular import by extracting shared helpers (#8980)
- Create api/voice_bundle_helpers.py to break circular dependency
- Move _require_admin() from voice_bundle_admin to helpers (GH#8980)
- Move _count_tools_for_bundle() from voice_bundle_user to helpers
- Update both modules to import from helpers instead of each other
- Fix test mocks to target correct module (api.voice_bundle_helpers)
- Resolves circular import: voice_bundle_user ↔ voice_bundle_admin

Previous attempt (418152ccd) imported directly causing ImportError.
This solution consolidates the admin check as requested while
maintaining clean module boundaries.
